Implement fullscreen on launch option
Add boolean option (defaults to false) to launch the fullscreen mode as
soon as the scorm content is started. We consider that the xblock has
"started" when we receive the first non-exit event. There is no other
way to proceed because we cannot detect click events inside the iframe.

This change also resolves bugs where the wrong xblock was updated when
modifying the grade or the fullscreen mode.