Add support for xblock completion
The scorm xblock did not have a green tick when completed. This is now
solved by emitting an event to the "completion" runtime service.

We decide to mark the xblock as completed when the user reaches the
final screen, aka depending on the cmi.core.lesson_status or
cmi.completion_status events. This means that the completion status
depends on the correct standard implementation by the scorm package.

Close #12.