Major performance improvements
Both GetValue and SetValue were being performed synchronously, which
caused clogging of the frontend with many get/set calls. To address this
we do the following:

- GetValue calls do not result in HTTP requests if the value is
available in the scorm_data.
- SetValue calls are performed asynchronously; however, the ordering of
the calls is preserved, for consistency. Values are stored in scorm_data
such that GetValue calls can return even faster.

These changes result in drastic user experience improvements.