fix: Problem with weight displaying related to localization is fixed
There is a problem with scorm weight displaying in settings from studio.
To reproduce it we need to open xblock settings menu, change the weight
value from default to new arbitrary one and save changes. After opening
this menu again the displayed weight value is empty. The reason is
localization. For the tested locale the float weight value is
represented with ',' instead of '.', but it is incorrect value for
input tag's `value` attribute, so nothing is displayed. To fix it,
`unlocalize` filter is applied to scorm_xblock.weight value.