fix: enable browser fullscreen mode for the scorm x-block
Enable browser fullscreen mode for the scorm x-block to enable correct rendering in the MFE application.

Fullscreen mode could not be enabled in the learning MFE because the xblock is rendered inside an iframe. As a consequence, we make use of the native fullscreen browser API to trigger actual fullscreen.

The legacy fullscreen document implementation is preserved for now to keep the "fullscreen on launch" feature, but it will be removed soon.